Create upperlowercapstring
Write a function called string_op. This function takes two arguments. The first one called string is a string on which an operation will be performed. The second one called command can be any variation of the words "upper", "lower", and "capitalize" (for example, "upper" could be spelled "uPPeR").

This function should return a new string that is the result of the command. If the command isn't valid, return "Invalid command!".