Create tup_hold_info
function called tup_hold_info that takes two arguments. The first, called lst, will hold a list of values. The second, called find, is the value to find. This function should return a tuple containing: the value that was supposed to be found, the number of times it was found, and a list of the indices where it was found. The values should be returned in the tuple in that order.